Add files via upload
This Flutter code creates a phone app that encourages people to throw away plastic responsibly. The main screen has sections explaining what the app is about, showing pictures, and guiding users through signing up and uploading photos of their plastic disposal. Users can earn points and redeem them for prizes. The app is designed to be easy to use and visually engaging, with interactive numbered circles. Overall, it aims to motivate people to contribute to reducing plastic waste and rewards them for doing so.